Fix issue with has_simd
